AK: Use size_t for CircularQueue and CircularDeque

+++ Applications/SystemMonitor/GraphWidget.cpp
@@ -58,7 +58,7 @@ void GraphWidget::paint_event(GUI::PaintEvent& event)
     float scale = (float)inner_rect.height() / (float)m_max;
 
     Gfx::Point prev_point;
-    for (int i = 0; i < m_values.size(); ++i) {
+    for (size_t i = 0; i < m_values.size(); ++i) {
         int x = inner_rect.right() - (i * 2) + 1;
         if (x < 0)
             break;
